+++ resolve/vector/VectorHandler.hpp
@@ -67,8 +67,8 @@ namespace ReSolve
               vector::Vector*     x,
               memory::MemorySpace memspace);
 
-    int elementwiseDivide(vector::Vector* divisor, vector::Vector* vec, memory::MemorySpace memspace);
-    int elementwiseMax(vector::Vector* x, vector::Vector* y, memory::MemorySpace memspace);
+    int scaleInv(vector::Vector* divisor, vector::Vector* vec, memory::MemorySpace memspace);
+    int max(vector::Vector* x, vector::Vector* y, memory::MemorySpace memspace);
 
     int abs(vector::Vector* x, memory::MemorySpace memspace);