+++ resolve/vector/VectorHandlerCuda.hpp
@@ -67,24 +67,24 @@ namespace ReSolve
     virtual void scal(vector::Vector* diag, vector::Vector* vec);
 
     /**
-     * @brief elementwiseDivide: divides a vector's elements by another's
+     * @brief scaleInv: divides a vector's elements by another's
      *
      * @param[in] divisor vector of size n x 1
      * @param[in,out] vec vector of size n x 1 (this is where the result is stored)
      *
      * @return 0 if successful, 1 otherwise
      */
-    virtual int elementwiseDivide(vector::Vector* divisor, vector::Vector* vec);
+    virtual int scaleInv(vector::Vector* divisor, vector::Vector* vec);
 
     /**
-     * @brief elementwiseMax: calculate the element-wise maximum of two vectors
+     * @brief max: calculate the element-wise maximum of two vectors
      *
      * @param[in]     x vector of size n x 1
      * @param[in,out] y vector of size n x 1 (this is where the result is stored)
      *
      * @return 0 if successful, 1 otherwise
      */
-    virtual int elementwiseMax(vector::Vector* x, vector::Vector* y);
+    virtual int max(vector::Vector* x, vector::Vector* y);
 
     /**
      * @brief abs: calculate the element-wise absolute value of a vector