+++ tests/unit/vector/runVectorHandlerTests.cpp
@@ -25,8 +25,8 @@ int main(int, char**)
     result += test.axpyMulti(100, 10);
     result += test.massDot(100, 10);
     result += test.scale(100);
-    result += test.elementwiseDivide(100);
-    result += test.elementwiseMax(100);
+    result += test.scaleInv(100);
+    result += test.max(100);
     result += test.abs(100);
 
     std::cout << "\n";
@@ -51,8 +51,8 @@ int main(int, char**)
     result += test.massDot(1000, 30);
     result += test.amax(1000);
     result += test.scale(1000);
-    result += test.elementwiseDivide(1000);
-    result += test.elementwiseMax(1000);
+    result += test.scaleInv(1000);
+    result += test.max(1000);
     result += test.abs(1000);
 
     std::cout << "\n";
@@ -78,8 +78,8 @@ int main(int, char**)
     result += test.massDot(1000, 30);
     result += test.amax(1000);
     result += test.scale(1000);
-    result += test.elementwiseDivide(1000);
-    result += test.elementwiseMax(1000);
+    result += test.scaleInv(1000);
+    result += test.max(1000);
     result += test.abs(1000);
 
     std::cout << "\n";